refactor user to member
