+++ pkg/core/resources/rbac/admin_resource_access_test.go
@@ -1,17 +1,17 @@
 package rbac_test
 
 import (
-	"github.com/kumahq/kuma/pkg/core/rbac"
-	resources_rbac "github.com/kumahq/kuma/pkg/core/resources/rbac"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
 	mesh_proto "github.com/kumahq/kuma/api/mesh/v1alpha1"
 	system_proto "github.com/kumahq/kuma/api/system/v1alpha1"
 	config_rbac "github.com/kumahq/kuma/pkg/config/rbac"
+	"github.com/kumahq/kuma/pkg/core/rbac"
 	"github.com/kumahq/kuma/pkg/core/resources/apis/mesh"
 	"github.com/kumahq/kuma/pkg/core/resources/apis/system"
 	"github.com/kumahq/kuma/pkg/core/resources/model"
+	resources_rbac "github.com/kumahq/kuma/pkg/core/resources/rbac"
 	"github.com/kumahq/kuma/pkg/core/user"
 )
 
@@ -55,7 +55,7 @@ var _ = Describe("Admin Resource Access", func() {
 		)
 
 		// then
-		Expect(err).To(MatchError(`access denied: user "john doe/users" of role "User" cannot access the resource of type "Secret"`))
+		Expect(err).To(MatchError(`access denied: user "john doe/users" of role "Member" cannot access the resource of type "Secret"`))
 	})
 
 	It("should deny anonymous user to access Create", func() {
@@ -94,7 +94,7 @@ var _ = Describe("Admin Resource Access", func() {
 		)
 
 		// then
-		Expect(err).To(MatchError(`access denied: user "john doe/users" of role "User" cannot access the resource of type "Secret"`))
+		Expect(err).To(MatchError(`access denied: user "john doe/users" of role "Member" cannot access the resource of type "Secret"`))
 	})
 
 	It("should allow admin to access Get", func() {
@@ -118,7 +118,7 @@ var _ = Describe("Admin Resource Access", func() {
 		)
 
 		// then
-		Expect(err).To(MatchError(`access denied: user "john doe/users" of role "User" cannot access the resource of type "Secret"`))
+		Expect(err).To(MatchError(`access denied: user "john doe/users" of role "Member" cannot access the resource of type "Secret"`))
 	})
 
 	It("should allow admin to access List", func() {
@@ -140,6 +140,6 @@ var _ = Describe("Admin Resource Access", func() {
 		)
 
 		// then
-		Expect(err).To(MatchError(`access denied: user "john doe/users" of role "User" cannot access the resource of type "Secret"`))
+		Expect(err).To(MatchError(`access denied: user "john doe/users" of role "Member" cannot access the resource of type "Secret"`))
 	})
 })
